backend/requirements.txt
fastapi
uvicorn
google-adk
python-dotenv
requests
aiosqlite
pydantic
pydantic-settings
